I added a navigation region for one of the maps
The navigation region only bakes when the tiles that it will use are in the first layer of the tile map. The wall is the second layer but it can still be used to make the navigation region if you set it to the first layer then put it back after the region has been baked.