Redid nav region for the first level for testing
I was hoping this would stop him from taking diagonal paths around corners because that always causes him to get stuck. This didn't have the effect i wanted, and it seems like the only real method to do what i want is to manually make each polygon, which I don't want to do.
My last idea is to make the "halls" in the pathfinding really narrow so that the ninja will only travel in the middle of the hall.